move icns logic out of react package
