+++ packages/icns/package.json
@@ -109,7 +109,9 @@
     "typescript": "4.8.3"
   },
   "dependencies": {
-    "@babel/runtime": "7.11.2"
+    "@babel/runtime": "7.11.2",
+    "@cosmos-kit/react": "^0.27.8",
+    "swr": "^2.0.0"
   },
   "peerDependencies": {
     "@cosmjs/amino": ">= 0.28",
